pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "neat-python>=0.92",
    "numpy>=2.2.6",
    "pygame>=2.6.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "icecream>=2.1.7",
]
